+++ lesson_17/bank/bank_app/src/main/java/com/codedifferently/lesson17/bank/CheckingAccount.java
@@ -1,6 +1,5 @@
 package com.codedifferently.lesson17.bank;
 
-import com.codedifferently.lesson17.bank.exceptions.InsufficientFundsException;
 import java.util.Set;
 
 /** Represents a checking account. */
@@ -15,24 +14,5 @@ public class CheckingAccount extends BankAccount {
    */
   public CheckingAccount(String accountNumber, Set<Customer> owners, double initialBalance) {
     super(accountNumber, owners, initialBalance);
-  }
-
-  /**
-   * Withdraws funds from the account.
-   *
-   * @param amount
-   * @throws InsufficientFundsException
-   */
-  public void withdraw(double amount) throws InsufficientFundsException {
-    if (isClosed()) {
-      throw new IllegalStateException("Cannot withdraw from a closed account");
-    }
-    if (amount <= 0) {
-      throw new IllegalStateException("Withdrawal amount must be positive");
-    }
-    if (balance < amount) {
-      throw new InsufficientFundsException("Account does not have enough funds for withdrawal");
-    }
-    balance -= amount;
   }
 }
