+++ lesson_17/bank/bank_app/src/main/java/com/codedifferently/lesson17/bank/SavingsAccount.java
@@ -1,6 +1,5 @@
 package com.codedifferently.lesson17.bank;
 
-import com.codedifferently.lesson17.bank.exceptions.InsufficientFundsException;
 import java.util.Set;
 
 /** Represents a savings account. */
@@ -15,15 +14,5 @@ public class SavingsAccount extends BankAccount {
    */
   public SavingsAccount(String accountNumber, Set<Customer> owners, double initialBalance) {
     super(accountNumber, owners, initialBalance);
-  }
-
-  /**
-   * Withdraws funds from the account.
-   *
-   * @param amount
-   * @throws InsufficientFundsException
-   */
-  public void withdraw(double amount) {
-    throw new UnsupportedOperationException("Withdrawals are not allowed from savings accounts");
   }
 }
